+++ src/Models/Configuration/Rules.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace PHPUnuhi\Models\Configuration;
+
+interface Rules
+{
+
+    /**
+     *
+     */
+    public const NESTING_DEPTH = 'nestingDepth';
+
+}