+++ tests/phpunit/Components/Validator/MixedStructureValidatorTest.php
@@ -95,6 +95,7 @@ class MixedStructureValidatorTest extends TestCase
             $locales,
             new Filter(),
             [],
+            [],
             []
         );
     }