+++ tests/phpunit/Models/Translation/TranslationSetTest.php
@@ -21,7 +21,7 @@ class TranslationSetTest extends TestCase
         $filter = new Filter();
         $locales = [];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $this->assertEquals('storefront', $set->getName());
     }
@@ -36,7 +36,7 @@ class TranslationSetTest extends TestCase
         $locales = [];
 
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $this->assertEquals('json', $set->getFormat());
     }
@@ -53,7 +53,7 @@ class TranslationSetTest extends TestCase
         $filter = new Filter();
         $locales = [];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $expected = [
             new Attribute('indent', '2'),
@@ -74,7 +74,7 @@ class TranslationSetTest extends TestCase
         $filter = new Filter();
         $locales = [];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $this->assertEquals('2', $set->getAttributeValue('indent'));
     }
@@ -90,7 +90,7 @@ class TranslationSetTest extends TestCase
         $filter = new Filter();
         $locales = [];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $this->assertEquals('', $set->getAttributeValue('abc'));
     }
@@ -107,7 +107,7 @@ class TranslationSetTest extends TestCase
         $locales[] = new Locale('', '', '');
         $locales[] = new Locale('', '', '');
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $this->assertCount(2, $set->getLocales());
     }
@@ -129,7 +129,7 @@ class TranslationSetTest extends TestCase
 
         $locales = [$localeEN, $localeDE];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $existing = $set->findAnyExistingTranslation('btnCancel');
 
@@ -153,7 +153,7 @@ class TranslationSetTest extends TestCase
         $filter = new Filter();
         $locales = [];
 
-        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, []);
+        $set = new TranslationSet('storefront', 'json', $locales, $filter, $attributes, [], []);
 
         $set->findAnyExistingTranslation('abc');
     }